Datasets refactoring
Extracted s3_location_handler
