+++ backend/dataall/modules/datasets/handlers/__init__.py
@@ -4,7 +4,9 @@ processing in a separate lambda function
 """
 from dataall.modules.datasets.handlers import (
     glue_column_handler,
-    glue_table_handler
+    glue_table_handler,
+    glue_profiling_handler,
+    s3_location_handler
 )
 
-__all__ = ["glue_column_handler", "glue_table_handler"]
+__all__ = ["glue_column_handler", "glue_table_handler", "glue_profiling_handler", "s3_location_handler"]
